+++ includes/extensions/class-menu.php
@@ -101,6 +101,46 @@ class WPS_Menu {
         <?php
     }
 
+    /**
+     * @param $item_id
+     * @param $menu_item
+     * @return void
+     */
+    public function addCustomFields($item_id, $menu_item)
+    {
+        $value = get_post_meta( $item_id, '_menu_item_aria_label', true );
+        ?>
+        <p class="field-aria-label description">
+            <label for="edit-menu-item-target-<?=$item_id?>>">
+                Aria label<br/>
+                <input type="text" id="edit-menu-item-target-<?=$item_id?>" class="widefat edit-menu-item-attr-title" value="<?php echo esc_attr($value); ?>" name="menu-item-aria-label[<?=$item_id?>]">
+            </label>
+        </p>
+        <?php
+    }
+
+
+    /**
+     * @param $menu_id
+     * @param $menu_item_db_id
+     * @return void
+     */
+    public function updateNavMenuItem($menu_id, $menu_item_db_id){
+
+        if ( isset( $_POST['menu-item-aria-label'][$menu_item_db_id]  ) ) {
+
+            $sanitized_data = sanitize_text_field( $_POST['menu-item-aria-label'][$menu_item_db_id] );
+            update_post_meta( $menu_item_db_id, '_menu_item_aria_label', $sanitized_data );
+
+        } else {
+
+            delete_post_meta( $menu_item_db_id, '_menu_item_aria_label' );
+        }
+    }
+
+    /**
+     * @return void
+     */
     public function addColumn()
     {
         global $wp_meta_boxes;
@@ -130,6 +170,8 @@ class WPS_Menu {
         $this->config = $_config;
 
         add_action( 'init', [$this, 'addMenus']);
+        add_action( 'wp_update_nav_menu_item', [$this, 'updateNavMenuItem'], 10, 2 );
         add_action( 'admin_head-nav-menus.php', [$this, 'addColumn']);
+        add_action( 'wp_nav_menu_item_custom_fields', [$this, 'addCustomFields'], 10, 2);
     }
 }